+++ src/index.js
@@ -1,0 +1,16 @@
+/**
+ * The main entry point of the application.
+ * 
+ * @module index
+ */
+
+import App from "./App";
+import './index.css';
+/**
+ * Create a new instance of the App class.
+ * 
+ * @type {App} - See {@link App}
+ */
+const app = new App();
+// Start the application
+app.run();